+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.3] - 2025-11-03
+## [0.0.4] - 2025-11-04
 
 ### Added 
 
-- WatchGuardRef.downcast_ref over Box<dyn Any>
+- WatchGuardRef::downcast over Box<dyn Any>
+- WatchGuardMut::downcast over Box<dyn Any>
 
 ## [0.0.2] - 2025-10-31